requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
datasets==1.18.2
dill==0.3.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
huggingface-hub==0.4.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
multidict==5.2.0
multiprocess==0.70.12.2
nltk==3.6.7
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pyarrow==6.0.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torchvision==0.11.2
tqdm==4.64.0
transformers==4.15.0
typing_extensions==4.1.1
urllib3==1.26.9
xxhash==3.0.0
yarl==1.7.2
zipp==3.6.0
